Move formatters to new file
